:art: Discount 똥 치우기 완료

+++ src/main/java/com/smplatform/product_service/domain/discount/service/impl/DiscountServiceImpl.java
@@ -3,15 +3,16 @@ package com.smplatform.product_service.domain.discount.service.impl;
 import com.smplatform.product_service.domain.discount.dto.DiscountRequestDto;
 import com.smplatform.product_service.domain.discount.dto.DiscountResponseDto;
 import com.smplatform.product_service.domain.discount.entity.Discount;
+import com.smplatform.product_service.domain.discount.exception.DiscountNotFoundException;
 import com.smplatform.product_service.domain.discount.repository.DiscountRepository;
 import com.smplatform.product_service.domain.discount.service.DiscountService;
-import com.smplatform.product_service.exception.DiscountNotFoundException;
 import lombok.RequiredArgsConstructor;
 import lombok.extern.slf4j.Slf4j;
 import org.springframework.stereotype.Service;
 
 import java.time.LocalDateTime;
 import java.util.ArrayList;
+import java.util.List;
 import java.util.Objects;
 import java.util.Optional;
 
@@ -24,27 +25,12 @@ public class DiscountServiceImpl implements DiscountService {
 
     // 할인코드 목록 조회
     @Override
-    public ArrayList<DiscountResponseDto.DiscountInfo> getDiscountList(LocalDateTime startDate, LocalDateTime endDate, String discountName) {
+    public List<DiscountResponseDto.DiscountInfo> getDiscountList(LocalDateTime startDate, LocalDateTime endDate, String discountName) {
+        List<Discount> discountEntityList = discountRepository.findDiscounts(startDate, endDate, discountName);
 
-        Optional<ArrayList<Discount>> discountList = Optional.of(new ArrayList<>());
-
-        if (Objects.isNull(discountName)) {
-            discountList = discountRepository.findByDiscountStartDateGreaterThanEqualAndDiscountEndDateLessThanEqual(startDate, endDate);
-        } else {
-            discountList = discountRepository.findByDiscountStartDateGreaterThanEqualAndDiscountEndDateLessThanEqualAndDiscountName(startDate, endDate, discountName);
-        }
-        if (discountList.isEmpty()) {
-            throw new DiscountNotFoundException();
-        }
-        ArrayList<Discount> discountEntityList = discountList.get();
-
-        ArrayList<DiscountResponseDto.DiscountInfo> discountInfos = new ArrayList<>();
-
-        for(Discount entity : discountEntityList) {
-            discountInfos.add(DiscountResponseDto.DiscountInfo.of(entity));
-        }
-
-        return discountInfos;
+        return discountEntityList.stream()
+                .map(DiscountResponseDto.DiscountInfo::of)
+                .toList();
     }
 
     // 할인코드 등록
@@ -62,7 +48,11 @@ public class DiscountServiceImpl implements DiscountService {
     @Override
     public String deleteDiscount(DiscountRequestDto.DeleteDiscount discountRequestDto) {
 
-        int discountId = discountRequestDto.toEntity().getDiscountId();
+        int discountId = discountRequestDto.getDiscountId();
+
+        if (!discountRepository.existsById(discountId)) {
+            throw new DiscountNotFoundException("삭제할 할인 코드가 존재하지 않습니다. ID: " + discountId);
+        }
 
         discountRepository.deleteById(discountId);
